refactor: restructure components by feature and add index barrels
- Organized components into feature-based folders: Explainer, Layout, Shared, Storyteller, Texts
- Renamed Common → Shared for clarity
- Added index.ts barrel files in each folder for cleaner imports
- Updated directory tree for better scalability
