+++ app/components/Explainer/ArticleDisplay.tsx
@@ -1,4 +1,4 @@
-import RenderMarkdown  from "~/components/ui/RenderMarkdown";
+import RenderMarkdown  from "~/components/Layout/RenderMarkdown";
 
 export function ArticleDisplay({ article }: { article: string }) {
   return (
