+++ app/components/Explainer/InstructionsBox.tsx
