+++ app/components/Shared/index.ts
@@ -1,0 +1,3 @@
+export * from './FadeIn';
+export * from './FeatureCard';
+export * from './SectionLoading';
