+++ app/components/Texts/VerseCard.tsx
