+++ app/components/Texts/index.ts
@@ -1,0 +1,2 @@
+export * from './VerseCard';
+export * from './VerseList';
