+++ app/routes/articles._index.tsx
@@ -3,8 +3,8 @@ import { Link, useLoaderData } from "@remix-run/react";
 import { Article } from "~/types";
 import type { LoaderFunction } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/node";
-import { PageFrame } from "~/components/ui/PageFrame";
-import { PageHeader } from "~/components/ui/PageHeader";
+import { PageFrame } from "~/components/Layout/PageFrame";
+import { PageHeader } from "~/components/Layout/PageHeader";
 import { getArticles } from "~/loader/articles";
 
 export const meta: MetaFunction = () => {
