+++ app/routes/learning._index.tsx
@@ -1,5 +1,5 @@
-import { PageFrame } from "~/components/ui/PageFrame";
-import { PageHeader } from "~/components/ui/PageHeader";
+import { PageFrame } from "~/components/Layout/PageFrame";
+import { PageHeader } from "~/components/Layout/PageHeader";
 export default function LearningIndex() {
   return (
     <PageFrame>
