+++ app/routes/texts.ribhu-gita.chapter-26.tsx
@@ -3,8 +3,8 @@ import type { MetaFunction } from "@remix-run/node";
 import { getRibhuGitaChapterData } from "~/loader/texts";
 import { useLoaderData } from "@remix-run/react";
 import { VerseList } from "~/components/VerseList";
-import { PageFrame } from "~/components/ui/PageFrame";
-import { PageHeader } from "~/components/ui/PageHeader";
+import { PageFrame } from "~/components/Layout/PageFrame";
+import { PageHeader } from "~/components/Layout/PageHeader";
 
 export const meta: MetaFunction = () => {
   return [
